pyproject.toml
[project]
name = "paper-decomposer"
version = "0.1.0"
description = "Paper decomposition pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf",
    "pyyaml",
    "pydantic>=2.0",
    "openai>=1.50",
    "rich",
    "tiktoken",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
    "api: marks tests that hit the Together API",
]
